requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apache-airflow==3.2.1
apache-airflow-core==3.2.1
apache-airflow-providers-celery==3.18.0
apache-airflow-providers-common-compat==1.14.3
apache-airflow-providers-common-io==1.7.2
apache-airflow-providers-common-sql==1.34.0
apache-airflow-providers-smtp==2.4.5
apache-airflow-providers-standard==1.12.3
apache-airflow-task-sdk==1.2.1
argcomplete==3.6.3
asgiref==3.11.1
attrs==26.1.0
babel==2.18.0
billiard==4.2.4
cadwyn==6.2.0
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.10.1
cron_descriptor==2.0.8
croniter==6.2.2
cryptography==46.0.7
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
fastapi-cli==0.0.24
flower==2.0.1
fsspec==2026.3.0
googleapis-common-protos==1.74.0
greenback==1.3.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.15.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.1.0
lockfile==0.12.2
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==11.0.2
msgspec==0.21.1
natsort==8.4.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
outcome==1.3.0.post0
packaging==26.1
pathlib_abc==0.5.2
pathspec==1.0.4
pendulum==3.2.0
pluggy==1.6.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
protobuf==6.33.6
psutil==7.2.2
pycparser==3.0
pydantic==2.13.1
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pygtrie==2.5.0
PyJWT==2.12.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.33.1
rich==13.9.4
rich-argparse==1.7.2
rich-toolkit==0.19.7
rpds-py==0.30.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==0.52.1
structlog==25.5.0
svcs==25.1.0
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
tornado==6.5.5
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uc-micro-py==2.0.0
universal_pathlib==0.3.10
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.44.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
wirerope==1.0.0
wrapt==2.1.2
zipp==3.23.1
google-cloud-bigquery==3.41.0
